rename hash to outgoing_message
