+++ cumulus/primitives/spec-messaging/src/lib.rs
@@ -21,4 +21,4 @@
 extern crate alloc;
 
 pub mod commitment_set;
-pub mod hash;
+pub mod outgoing_message;
